修复bug

+++ QuChu/app/src/main/java/co/quchu/quchu/model/FavoriteModel.java
@@ -71,7 +71,7 @@ public class FavoriteModel {
             private int id;
             private String name;
             private String rgb;
-            private int score;
+            private float score;
 
             public void setCover(String cover) {
                 this.cover = cover;
@@ -89,7 +89,7 @@ public class FavoriteModel {
                 this.rgb = rgb;
             }
 
-            public void setScore(int score) {
+            public void setScore(float score) {
                 this.score = score;
             }
 
@@ -109,7 +109,7 @@ public class FavoriteModel {
                 return rgb;
             }
 
-            public int getScore() {
+            public float getScore() {
                 return score;
             }
         }
